Split Noosphere classes into separate files
Additionally, make SphereProtocol.read a throwing function so we can
gain more insight into failures.
