+++ xcode/Subconscious/Shared/Services/Noosphere/Noosphere.swift
@@ -1,0 +1,225 @@
+//
+//  Noosphere.swift
+//  Subconscious (iOS)
+//
+//  Created by Gordon Brander on 2/13/23.
+//
+//  Noosphere lifecycle:
+//  1. Initialize namespace (once ever)
+//  2. Create a unique key (once per key)
+//  3. Create sphere for key, getting back a receipt (once per key)
+//     - Retrieve identity from sphere receipt, then store (once, upon sphere creation)
+//     - Retrieve sphere mnemonic from sphere receipt. Display to user, then discard (once, upon sphere creation)
+//  4. Open sphere file system (on-demand)
+
+import Foundation
+import SwiftNoosphere
+
+public enum NoosphereError: Error, LocalizedError {
+    /// Thrown when something unexpected happens on the other side of the FFI, and we don't know what went wrong.
+    case foreignError(String)
+    /// Thrown when an OpaquePointer? is unwrapped and found to be a null pointer.
+    case nullPointer
+    case contentTypeMissing(_ slashlink: String)
+    case fileDoesNotExist(_ slashlink: String)
+    
+    public var errorDescription: String? {
+        switch self {
+        case .foreignError(let message):
+            return "Foreign Error: \(message)"
+        case .nullPointer:
+            return "Null pointer"
+        case .contentTypeMissing(let slashlink):
+            return "Content-Type header is missing for file: \(slashlink)"
+        case .fileDoesNotExist(let slashlink):
+            return "File does not exist: \(slashlink)"
+        }
+    }
+}
+
+struct NoosphereFFI {
+    static func callWithError<Z>(
+        _ perform: (UnsafeMutablePointer<OpaquePointer?>) -> Z
+    ) throws -> Z {
+        let error = UnsafeMutablePointer<OpaquePointer?>.allocate(capacity: 1)
+        defer {
+            error.deallocate()
+        }
+        let value = perform(error)
+        if let errorPointer = error.pointee {
+            defer {
+                ns_error_free(errorPointer)
+            }
+            guard let errorMessagePointer = ns_error_string(
+                errorPointer
+            ) else {
+                throw NoosphereError.foreignError("Unknown")
+            }
+            defer {
+                ns_string_free(errorMessagePointer)
+            }
+            let errorMessage = String.init(cString: errorMessagePointer)
+            throw NoosphereError.foreignError(errorMessage)
+        }
+        return value
+    }
+    
+    static func callWithError<A, Z>(
+        _ perform: (A, UnsafeMutablePointer<OpaquePointer?>) -> Z,
+        _ a: A
+    ) throws -> Z {
+        try Self.callWithError { error in perform(a, error) }
+    }
+    
+    static func callWithError<A, B, Z>(
+        _ perform: (A, B, UnsafeMutablePointer<OpaquePointer?>) -> Z,
+        _ a: A,
+        _ b: B
+    ) throws -> Z {
+        try Self.callWithError { error in perform(a, b, error) }
+    }
+    
+    static func callWithError<A, B, C, Z>(
+        _ perform: (A, B, C, UnsafeMutablePointer<OpaquePointer?>) -> Z,
+        _ a: A,
+        _ b: B,
+        _ c: C
+    ) throws -> Z {
+        try Self.callWithError { error in perform(a, b, c, error) }
+    }
+
+    /// Read first header value for file pointer
+    static func readFileHeaderValueFirst(
+        file: OpaquePointer,
+        name: String
+    ) -> String? {
+        guard let valueRaw = ns_sphere_file_header_value_first(
+            file,
+            name
+        ) else {
+            return nil
+        }
+        defer {
+            ns_string_free(valueRaw)
+        }
+        return String(cString: valueRaw)
+    }
+
+    /// Get all header names for a given file pointer
+    static func readFileHeaderNames(
+        file: OpaquePointer
+    ) -> [String] {
+        let file_header_names = ns_sphere_file_header_names_read(file)
+        defer {
+            ns_string_array_free(file_header_names)
+        }
+
+        let name_count = file_header_names.len
+        guard var pointer = file_header_names.ptr else {
+            return []
+        }
+
+        var names: [String] = []
+        for _ in 0..<name_count {
+            let name = String(cString: pointer.pointee!)
+            names.append(name)
+            pointer += 1;
+        }
+        return names
+    }
+}
+
+public struct SphereReceipt {
+    var identity: String
+    var mnemonic: String
+}
+
+/// Create a Noosphere instance.
+///
+/// - Property noosphere: pointer that holds all the internal book keeping.
+///   DB pointers, key storage interfaces, active HTTP clients etc.
+public final class Noosphere {
+    let noosphere: OpaquePointer
+    
+    init(
+        globalStoragePath: String,
+        sphereStoragePath: String,
+        gatewayURL: String? = nil
+    ) throws {
+        guard let noosphere = try NoosphereFFI.callWithError(
+            ns_initialize,
+            globalStoragePath,
+            sphereStoragePath,
+            gatewayURL
+        ) else {
+            throw NoosphereError.nullPointer
+        }
+        self.noosphere = noosphere
+    }
+    
+    /// Create and configure a user and sphere
+    /// - Important: the identity should be persisted
+    /// - Important: the mnemonic should be displayed to user and then discarded.
+    /// - Returns: a SphereConfig containing the sphere identity and mnemonic
+    ///
+    /// What it does:
+    /// - Initializes a namespace
+    /// - Creates a key
+    /// - Creates a sphere for that key
+    public func createSphere(
+        ownerKeyName: String
+    ) throws -> SphereReceipt {
+        try NoosphereFFI.callWithError(
+            ns_key_create,
+            noosphere,
+            ownerKeyName
+        )
+        
+        guard let sphereReceipt = try NoosphereFFI.callWithError(
+            ns_sphere_create,
+            noosphere,
+            ownerKeyName
+        ) else {
+            throw NoosphereError.nullPointer
+        }
+        defer {
+            ns_sphere_receipt_free(sphereReceipt)
+        }
+        
+        guard let sphereIdentityPointer = try NoosphereFFI.callWithError(
+            ns_sphere_receipt_identity,
+            sphereReceipt
+        ) else {
+            throw NoosphereError.foreignError(
+                "Failed to get pointer for identity"
+            )
+        }
+        defer {
+            ns_string_free(sphereIdentityPointer)
+        }
+        
+        guard let sphereMnemonicPointer = try NoosphereFFI.callWithError(
+            ns_sphere_receipt_mnemonic,
+            sphereReceipt
+        ) else {
+            throw NoosphereError.foreignError(
+                "Failed to get pointer for mnemonic"
+            )
+        }
+        defer {
+            ns_string_free(sphereMnemonicPointer)
+        }
+        
+        let sphereIdentity = String.init(cString: sphereIdentityPointer)
+        let sphereMnemonic = String.init(cString: sphereMnemonicPointer)
+        
+        return SphereReceipt(
+            identity: sphereIdentity,
+            mnemonic: sphereMnemonic
+        )
+    }
+    
+    deinit {
+        ns_free(noosphere)
+    }
+}
