+++ xcode/Subconscious/Shared/Services/Noosphere/SphereFS.swift
@@ -1,0 +1,331 @@
+//
+//  SphereFS.swift
+//  Subconscious (iOS)
+//
+//  Created by Gordon Brander on 2/13/23.
+//
+
+import Foundation
+import SwiftNoosphere
+
+/// Describes a Sphere.
+/// See `SphereFS` for a concrete implementation.
+public protocol SphereProtocol {
+    associatedtype Memo
+    
+    func version() throws -> String
+    
+    func getFileVersion(slashlink: String) -> String?
+    
+    func readHeaderValueFirst(
+        slashlink: String,
+        name: String
+    ) -> String?
+    
+    func readHeaderNames(slashlink: String) -> [String]
+    
+    func read(slashlink: String) throws -> Memo
+    
+    func write(
+        slug: String,
+        contentType: String,
+        additionalHeaders: [Header],
+        body: Data
+    ) throws
+    
+    func remove(slug: String) throws
+    
+    func save() throws -> String
+    
+    func list() throws -> [String]
+    
+    func sync() throws -> String
+    
+    func changes(_ since: String?) throws -> [String]
+}
+
+/// Sphere file system access.
+/// Provides sphere file system methods and manages lifetime of sphere pointer.
+public final class SphereFS: SphereProtocol {
+    private let noosphere: Noosphere
+    public let fs: OpaquePointer
+    public let identity: String
+    
+    init(noosphere: Noosphere, identity: String) throws {
+        self.noosphere = noosphere
+        self.identity = identity
+        guard let fs = try NoosphereFFI.callWithError(
+            ns_sphere_fs_open,
+            noosphere.noosphere,
+            identity
+        ) else {
+            throw NoosphereError.nullPointer
+        }
+        self.fs = fs
+    }
+    
+    /// Get current version of sphere
+    public func version() throws -> String {
+        guard let sphereVersionPointer = try NoosphereFFI.callWithError(
+            ns_sphere_version_get,
+            noosphere.noosphere,
+            identity
+        ) else {
+            throw NoosphereError.nullPointer
+        }
+        defer {
+            ns_string_free(sphereVersionPointer)
+        }
+        return String.init(cString: sphereVersionPointer)
+    }
+
+    /// Read first header value for memo at slashlink
+    /// - Returns: value, if any
+    public func readHeaderValueFirst(
+        slashlink: String,
+        name: String
+    ) -> String? {
+        guard let file = try? NoosphereFFI.callWithError(
+            ns_sphere_fs_read,
+            noosphere.noosphere,
+            fs,
+            slashlink
+        ) else {
+            return nil
+        }
+        defer {
+            ns_sphere_file_free(file)
+        }
+        return NoosphereFFI.readFileHeaderValueFirst(
+            file: file,
+            name: name
+        )
+    }
+    
+    /// Get the base64-encoded CID v1 string for the memo that refers to the
+    /// content of this sphere file.
+    public func getFileVersion(slashlink: String) -> String? {
+        guard let file = try? NoosphereFFI.callWithError(
+            ns_sphere_fs_read,
+            noosphere.noosphere,
+            fs,
+            slashlink
+        ) else {
+            return nil
+        }
+        guard let cid = try? NoosphereFFI.callWithError(
+            ns_sphere_file_version_get,
+            file
+        ) else {
+            return nil
+        }
+        defer {
+            ns_string_free(cid)
+        }
+        return String.init(cString: cid)
+    }
+
+    /// Read all header names for a given slashlink
+    public func readHeaderNames(slashlink: String) -> [String] {
+        guard let file = try? NoosphereFFI.callWithError(
+            ns_sphere_fs_read,
+            noosphere.noosphere,
+            fs,
+            slashlink
+        ) else {
+            return []
+        }
+        defer {
+            ns_sphere_file_free(file)
+        }
+        return NoosphereFFI.readFileHeaderNames(file: file)
+    }
+    
+    /// Read the value of a memo from a Sphere
+    /// - Returns: `Memo`
+    public func read(slashlink: String) throws -> MemoData {
+        guard let file = try NoosphereFFI.callWithError(
+            ns_sphere_fs_read,
+            noosphere.noosphere,
+            fs,
+            slashlink
+        ) else {
+            throw NoosphereError.fileDoesNotExist(slashlink)
+        }
+        defer {
+            ns_sphere_file_free(file)
+        }
+        
+        guard let contentType = NoosphereFFI.readFileHeaderValueFirst(
+            file: file,
+            name: "Content-Type"
+        ) else {
+            throw NoosphereError.contentTypeMissing(slashlink)
+        }
+        
+        let bodyRaw = try NoosphereFFI.callWithError(
+            ns_sphere_file_contents_read,
+            noosphere.noosphere, file
+        )
+        defer {
+            ns_bytes_free(bodyRaw)
+        }
+        let body = Data(bytes: bodyRaw.ptr, count: bodyRaw.len)
+        
+        var headers: [Header] = []
+        let headerNames = NoosphereFFI.readFileHeaderNames(file: file)
+        for name in headerNames {
+            // Skip content type. We've already retreived it.
+            guard name != "Content-Type" else {
+                continue
+            }
+            guard let value = NoosphereFFI.readFileHeaderValueFirst(
+                file: file,
+                name: name
+            ) else {
+                continue
+            }
+            headers.append(Header(name: name, value: value))
+        }
+        
+        return MemoData(
+            contentType: contentType,
+            additionalHeaders: headers,
+            body: body
+        )
+    }
+    
+    /// Write to sphere
+    public func write(
+        slug: String,
+        contentType: String,
+        additionalHeaders: [Header] = [],
+        body: Data
+    ) throws {
+        try body.withUnsafeBytes({ rawBufferPointer in
+            let bufferPointer = rawBufferPointer.bindMemory(to: UInt8.self)
+            let pointer = bufferPointer.baseAddress!
+            let bodyRaw = slice_ref_uint8(
+                ptr: pointer, len: body.count
+            )
+            
+            guard let additionalHeadersContainer = ns_headers_create() else {
+                throw NoosphereError.foreignError("ns_headers_create failed to return pointer")
+            }
+            defer {
+                ns_headers_free(additionalHeadersContainer)
+            }
+            
+            for header in additionalHeaders {
+                ns_headers_add(
+                    additionalHeadersContainer,
+                    header.name,
+                    header.value
+                )
+            }
+            
+            try NoosphereFFI.callWithError { error in
+                ns_sphere_fs_write(
+                    noosphere.noosphere,
+                    fs,
+                    slug,
+                    contentType,
+                    bodyRaw,
+                    additionalHeadersContainer,
+                    error
+                )
+            }
+        })
+    }
+
+    /// Save outstanding writes and return new Sphere version
+    @discardableResult public func save() throws -> String {
+        try NoosphereFFI.callWithError(
+            ns_sphere_fs_save,
+            noosphere.noosphere,
+            fs,
+            nil
+        )
+        return try self.version()
+    }
+    
+    /// Remove slug from sphere
+    public func remove(slug: String) throws {
+        try NoosphereFFI.callWithError(
+            ns_sphere_fs_remove,
+            noosphere.noosphere,
+            fs,
+            slug
+        )
+    }
+    
+    /// List all slugs in sphere
+    public func list() throws -> [String] {
+        let slugs = try NoosphereFFI.callWithError(
+            ns_sphere_fs_list,
+            noosphere.noosphere,
+            fs
+        )
+        defer {
+            ns_string_array_free(slugs)
+        }
+
+        let slugCount = slugs.len
+        var pointer = slugs.ptr!
+
+        var output: [String] = []
+        for _ in 0..<slugCount {
+            let slug = String.init(cString: pointer.pointee!)
+            output.append(slug)
+            pointer += 1;
+        }
+        return output
+    }
+
+    /// Sync sphere with gateway.
+    /// Gateway must be configured when Noosphere was initialized.
+    public func sync() throws -> String {
+        let versionPointer = try NoosphereFFI.callWithError(
+            ns_sphere_sync,
+            noosphere.noosphere,
+            identity
+        )
+        guard let versionPointer = versionPointer else {
+            throw NoosphereError.nullPointer
+        }
+        defer {
+            ns_string_free(versionPointer)
+        }
+        return String(cString: versionPointer)
+    }
+    
+    /// List all changed slugs between two versions of a sphere.
+    /// This method lists which slugs changed between version, but not
+    /// what changed.
+    public func changes(_ since: String? = nil) throws -> [String] {
+        let changes = try NoosphereFFI.callWithError(
+            ns_sphere_fs_changes,
+            noosphere.noosphere,
+            fs,
+            since
+        )
+        defer {
+            ns_string_array_free(changes)
+        }
+
+        let changesCount = changes.len
+        var pointer = changes.ptr!
+        
+        var slugs: [String] = []
+        for _ in 0..<changesCount {
+            let slug = String.init(cString: pointer.pointee!)
+            slugs.append(slug)
+            pointer += 1;
+        }
+        return slugs
+    }
+
+    deinit {
+        ns_sphere_fs_free(fs)
+    }
+}
